Fixed lag while scrolling (Read description if curious) + other small things
- Lag on team page -
There were mainly 2 CSS properties making this page extremely laggy, background-filter: blur(10px) (about 80% of the lag) and opacity: 0.7 (about 20% of the lag). It seems the use of these effects used on a large scale is too much for the browser to process.

- Lag on home page -
background-attachment: fixed- This for some reason causes some lag on the languages section. It might have something to do with the semi-transparent background color above it, idk.